Moving client configuration options from realis to its own sepearate file to make code more digestible.
